+++ src/main/java/com/perz/carrentalapp/model/User.java
@@ -1,7 +1,5 @@
 package com.perz.carrentalapp.model;
 
-
-import com.perz.carrentalapp.model.Role;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -51,16 +49,4 @@ public class User {
     @Column(name ="disabled")
     private Boolean disabled;
 
-    @Override
-    public String toString() {
-        return "User{" +
-                "id=" + id +
-                ", firstname='" + firstname + '\'' +
-                ", lastname='" + lastname + '\'' +
-                ", email='" + email + '\'' +
-                ", phone='" + phone + '\'' +
-                ", password='" + password + '\'' +
-                ", role=" + role +
-                '}';
-    }
 }
