+++ src/main/java/com/perz/carrentalapp/repositories/UserRepository.java
@@ -1,6 +1,7 @@
 package com.perz.carrentalapp.repositories;
 
 
+
 import com.perz.carrentalapp.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
