+++ src/main/java/com/perz/carrentalapp/util/Converter.java
@@ -1,7 +1,9 @@
 package com.perz.carrentalapp.util;
 
 import com.perz.carrentalapp.model.User;
-import com.perz.carrentalapp.auth.dto.UserRegistrationDTO;
+import com.perz.carrentalapp.model.dto.UserDTO;
+import com.perz.carrentalapp.model.dto.UserRegistrationDTO;
+import com.perz.carrentalapp.model.dto.UserToUpdateDTO;
 import org.modelmapper.ModelMapper;
 
 
@@ -9,8 +11,15 @@ public class Converter {
 
     private final static ModelMapper modelMapper = new ModelMapper();
 
-    public static User convertFromUserDTOToUser(UserRegistrationDTO userDTO) {
+    public static User convertFromUserRegistrationDTOToUser(UserRegistrationDTO userDTO) {
         return modelMapper.map(userDTO,User.class);
     }
 
+    public static UserDTO convertFromUserToUserDTO(User user) {
+        return modelMapper.map(user, UserDTO.class);
+    }
+
+    public static User convertFromUserToUpdateDTOToUser(UserToUpdateDTO userToUpdateDTO) {
+        return modelMapper.map(userToUpdateDTO,User.class);
+    }
 }
