+++ src/main/java/com/perz/carrentalapp/util/UserErrorResponse.java
@@ -1,4 +1,4 @@
-package com.perz.carrentalapp.auth.util;
+package com.perz.carrentalapp.util;
 
 import lombok.Getter;
 import lombok.Setter;
